RS-4292 Updates to allow specs to run in 5.1
Adds version numbers to the migration files in line with when they were created.

Adds required: false to belongs_to associations which are not required and
updates those which are required to be explicit so we do not rely on the config
being set up properly in an app.

Removes a spec which doesn't run any more due to removed assets as the FE
concerns were removed.